Explicitly deploy autoretrive pods to `dev-ue2a-r5a-2xl` node group
The autoretieve deployment was configured to use up to 32Gi of memory.
Currently, autoretrieve gets rescheduled because it consumes more memory
than it is allowed to every 12 hours or so. The worker node group set up
for it however has up to 64Gi memory. This means when autoretrieve is
scheduled, there is spare capacity on the worker node to run other pods
which could delay the scheduling of autoretrieve when it is restarted.

Considering the memory consumption of autoretrieve, increase the memory
limit to 60Gi so that there is more memory available to the pod while
leaving 4Gi of spare memory for the other administrative pods that need
to run on every node (like log forwarders etc.).

Additionally, explicitly configure node selectors to have a predictable
scheduling for autorerive pods on worker node group created for it.

The combination of these two changes mean: autoretrieve will always get
scheduled on the AZ where its current PVC is, and because it demands
larger memory there will be no spare capacity to run other pods that
could delay its startup.

Note, the resouce limit config is moved from `base` overlay to `dev`
since it is a dev-specific configuration.